Initial Commit
Contains a full dictionary of "hello" messages, a partial of "Merry Christmas" greetings, and a handful of good outcome messages. Needs to be refactored to classes are represented in different files. Will probably make a message class so that more information regarding the language can be added in a "nice" way. So message, language, language code, those types of properties.